pyproject.toml
[tool.poetry]
name = "wallpaperbot"
version = "0.1.0"
description = "Cross-platform wallpaper downloader/manager for triple-monitor setups"
authors = ["digitalsorc <ops@digitalsorc.dev>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "wallpaperbot" }]

[tool.poetry.dependencies]
python = "^3.12"
typer = { version = "^0.12.5", extras = ["all"] }
httpx = "^0.27.2"
tenacity = "^9.0.0"
pyyaml = "^6.0.2"
pydantic = "^2.9.2"
Pillow = "^11.0.0"
imagehash = "^4.3.1"
structlog = "^24.4.0"
rich = "^13.9.2"
platformdirs = "^4.3.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
respx = "^0.21.1"
pytest-mock = "^3.14.0"
freezegun = "^1.5.1"
ruff = "^0.7.0"

[tool.poetry.scripts]
wallpaperbot = "wallpaperbot.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = ["E", "F", "I", "B", "UP", "SIM", "ASYNC"]
lint.ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"